requirements.txt
accelerate==1.6.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
auto-mix-prep==0.2.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.1
docling-core==2.26.1
docutils==0.21.2
exceptiongroup==1.2.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
gitpython==3.1.44
gliner==0.2.17
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langid==1.1.6
latex2mathml==3.77.0
marimo==0.12.8
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.34.1
networkx==3.4.2
numpy==2.1.1
onnxruntime==1.21.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycrdt==0.11.1
pydantic==2.11.3
pydantic-core==2.33.1
pydeck==0.9.1
pydub==0.25.1
pygments==2.19.1
pymdown-extensions==10.14.3
pymupdf==1.25.5
pyparsing==3.2.3
pypdf2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
ruff==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
st-diff-viewer==1.0.7
starlette==0.46.1
streamlit==1.44.1
streamlit-pdf-viewer==0.0.22
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.4.1+cpu
torchvision==0.19.1+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.1
typer==0.12.5
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
websockets==15.0.1
